requirements.txt
APScheduler==3.10.1
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.14.0
djangorestframework==3.13.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.7
tzlocal==4.3